+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+## [2.0.11] - 2024-08-18
+
 ### Added
 
 - Completion for optional paths in projects and workspace config
@@ -121,7 +123,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Initial release
 
-[unreleased]: https://github.com/argon-rbx/argon-vscode/compare/2.0.10...HEAD
+[unreleased]: https://github.com/argon-rbx/argon-vscode/compare/2.0.11...HEAD
+[2.0.11]: https://github.com/argon-rbx/argon-vscode/compare/2.0.10...2.0.11
 [2.0.10]: https://github.com/argon-rbx/argon-vscode/compare/2.0.9...2.0.10
 [2.0.9]: https://github.com/argon-rbx/argon-vscode/compare/2.0.8...2.0.9
 [2.0.8]: https://github.com/argon-rbx/argon-vscode/compare/2.0.7...2.0.8